Keep local hour across daylight saving time change (#27)
* Keep local hour across daylight saving time change

For recurring type, the next schedule time does not keep the same local time hour if the current time offset is a different offset than the initial setup time due to daylight time saving change. This change uses local time to determine the next schedule day.

The schedule entry is updated with the next date/time to always use the current time change policy, to eliminate overhead of calculating it each reference and so can view in the schedule entry. The initial date/time is not retained. For tracking, the schedule date/time update is logged.

To verify the calculation, set a schedule with the date at least 2 days before the previous daylight saving time change. Refresh the list. The schedule will have the next schedule date with the same local hour. The logic does not work for a schedule for only the hour of the time change where the offset increases (daylight saving time to standard time).

* Use DateTime

Newer DateTimeImmutable is not needed.  I did not realize older PHP was still supported.